refactor: reorganize project structure and improve modularity
- Split models into dedicated stock.py and portfolio.py
- Split utilities into market.py and formatting.py
- Improve separation of concerns in all modules
- Update imports to reflect new structure
- Enhance code documentation
- Maintain consistent directory organization
