+++ utils/__init__.py
@@ -1,1 +1,24 @@
-"""Utilidades para el manejo de stocks y portfolios."""
+"""Utility functions for the stocks portfolio package."""
+
+from .formatting import format_currency, format_percentage, print_logo, print_stock_info
+from .market import (
+    get_next_trading_day,
+    get_stock_history,
+    get_stock_price,
+    get_ticker,
+    is_trading_day,
+    validate_symbol,
+)
+
+__all__ = [
+    "format_currency",
+    "format_percentage",
+    "print_logo",
+    "print_stock_info",
+    "get_next_trading_day",
+    "get_stock_history",
+    "get_stock_price",
+    "get_ticker",
+    "is_trading_day",
+    "validate_symbol",
+]
